+++ doc/sphinx/src/news.rst
@@ -15,6 +15,16 @@ Release Notes
 
    The bullets in each such discussion are in descending order of importance.
 
+Changes in v2.5.0
++++++++++++++++++
+
+New features
+^^^^^^^^^^^^
+* Execution
+
+  * ``params_tag`` is a base class for user-defined classes that are a composition of task-parameter types (like field accessors).
+  * ``arg_tag`` is a base class for user-defined classes that specifiy custom handling as a task argument.
+
 Changes in v2.4.1
 +++++++++++++++++